% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{ClaritySim-package}
\alias{ClaritySim}
\alias{ClaritySim-package}
\title{(Sim)ulate (Sim)ilarity between (Sim)ilarity Matrices}
\description{
Simulate Similarity Matrices for comparison and use in Packages \code{\link[SimSim]{SimSim}}. This package generates random trees, random mixtures under those trees. It then allows randomly purturbing either the branch length or the topology of the tree, all the while reporting the similarity between objects which are themselves mixtures of tips of the tree.
}
\references{
"Modelling Similarity Matrices as Mixtures for Structural Comparison" by Daniel Lawson, 2018 (dan.lawson@bristol.ac.uk)
}
\author{
\strong{Maintainer}: Daniel Lawson \email{dan.lawson@bristol.ac.uk}

}
