% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClaritySim.R
\name{resimulatedDistances}
\alias{resimulatedDistances}
\title{Distance between a simulation and replicates of itself with the same structure}
\usage{
resimulatedDistances(
  sim,
  nrep = 100,
  distfn = function(x) as.matrix(stats::dist(x))
)
}
\arguments{
\item{sim}{A ClaritySim object as returned by \code{simulateCoalescent}}

\item{nrep}{(default=100) number of replications to average over}

\item{distfn}{(default=as.matrix(stats::dist(x))) distance function to compute}
}
\value{
A ClaritySim object with an updated Y
}
\description{
Computes the euclidean distance between all points by replicating the data x to x', and reporting the distance between each x and x' pair. This gives a fair representation of the diagonal.

See \code{c_dist} in package \code{Clarity} which implements a simpler solution to the diagonal.
}
\examples{
\dontrun{
set.seed(1)
mysim=simulateCoalescent(36,36,L=200,
                         alpha=rep(0,36),sigma0=0.1)
mysim2a<-resimulatedDistances(mysim,nrep=500)
mycols=matrix(1,nrow=dim(mysim$Y)[1],ncol=dim(mysim$Y)[2])
diag(mycols)=2

plot((mysim$Y),(mysim2a$Y),
     xlab="default: minimum from each row",
     ylab="better: simulationToDistance",col=mycols)
abline(a=0,b=1)
legend("topleft",legend=c("off-diagonal","diagonal"),
       text.col=1:2,bty="n")
}
}
\seealso{
\code{\link{claritysim_dist}}, which does it the fast way as recommended in Clarity.
}
