% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClaritySim.R
\name{simData}
\alias{simData}
\title{Simulate features under a ClaritySim "mixture on a tree model"}
\usage{
simData(sim, L = NULL, sigma = NULL, sigma0 = NULL)
}
\arguments{
\item{sim}{a ClaritySim simulation containing a tree}

\item{L}{(default NULL meaning take from sim) number of features to simulate}

\item{sigma}{(default NULL meaning take from sim) the rate of drift of features per unit coalescence distance}

\item{sigma0}{(default NULL meaning take from sim) the 'individual-level noise' in the trait values after accounting for individual mixtures}
}
\value{
A ClaritySim object, a list with updated values of
\itemize{
\item D0, the trait values associated with the tips of each branch of the tree
\item D, the trait values associated with each data point
\item Y, the matrix of distances for the samples
\item L, as input
\item sigma, as input
\item sigma0, as input
}
}
\description{
Simulate traits assuming random walk using the cophenetic imposed using the tree. Specifically, data samples are generated that contain drift according to the tree, then a value created by their average mixture from that tree, plus noise
}
\seealso{
\code{\link{simulateCoalescent}}
}
\keyword{mixture}
