% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClaritySim.R
\name{transformCoalescent}
\alias{transformCoalescent}
\title{rescale similarities of a simulated coalescent to new branch lengths}
\usage{
transformCoalescent(sim, multmin = 0.1, multmax = 2)
}
\arguments{
\item{sim}{A simulated coalescent as returned by \code{\link{simulateCoalescent}}}

\item{multmin}{(default 0.1) minimum multiplier for branch edge lengths}

\item{multmax}{(default 2) maximum multiplier for branch edge lengths}
}
\value{
A list containing the same objects as  \code{\link{simulateCoalescent}} with updated X, Y, Y0, tree elements
}
\description{
Take a simulated tree and multiply each edge length by a random amount to generate a tree with the same topology but different structure.
}
\examples{
set.seed(1)
mysim<-simulateCoalescent(100,10) # Simlulate 100 objects in a 10 dimensional latent space
similarsim<-transformCoalescent(mysim)
# similarsim$Y is the data datarep from package Clarity

}
\seealso{
\code{\link{simulateCoalescent}}, \code{\link{mixCoalescent}}
}
\keyword{mixture}
