% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClaritySim.R
\name{simulateCoalescent}
\alias{simulateCoalescent}
\title{simulate similarities under a coalescent model}
\usage{
simulateCoalescent(N, K, alpha = rep(0.2, K), sigma0 = 0.01)
}
\arguments{
\item{N}{Number of samples}

\item{K}{Dimension of the latent structure}

\item{alpha}{(default rep(1.5,K) ) structure of calX}

\item{sigma0}{(default 0.1) Noise in the vector (Can be a vector or a scalar)}
}
\value{
A list containing:
\itemize{
\item Y, an N by N non-negative matrix of similarities
\item Y0, Y but without the noise added
\item A, an N by K non-negative matrix of mixtures
\item X, a K by K non-negative matrix of similarities between latent objects
\item calX, a K by K non-negative matrix of similarities between naturally scaled latent objects
\item tree, the tree that was simulated in \code{\link{rcoal}} output format
\item sigma0 as input
item alpha as input
}
}
\description{
Simulate a coalescent relationship between K latent objects. Then simulate N samples of those objects as dirichlet-distriubted mixtures. Finally, generate a latent structure (a K by K matrix) calX which relates the latent objects to one another using cophenetic distance on the tree, and map this into the data to create the N by N similarity matrix Y (with noise)
}
\examples{
set.seed(1)
mysim<-simulateCoalescent(100,10) # Simlulate 100 objects in a 10 dimensional latent space
# This is the data dataraw from Package Clarity

}
\seealso{
\code{\link{mixCoalescent}}, \code{\link{transformCoalescent}}
}
\keyword{mixture}
